SparseBase

Documentation for SparseBase.

SparseBase.filltype

SparseBase.fkeep!

SparseBase.getfill

SparseBase.hasfixedsparsity

SparseBase.haszerobasedinternals

SparseBase.isisovalued

SparseBase.isopaque

SparseBase.issparse

SparseBase.mapstored!

SparseBase.nstored

SparseBase.setfill

SparseBase.storedindices

SparseBase.storedvalues

SparseBase.filltype — Method

filltype(A)

Type of implicit values of A. Most arrays either have no fill, or only support fill in the same domain as eltype(A).

SparseBase.fkeep! — Function

fkeep!(f, [C], A)

Keep values of A where f is true. f(i..., x)::Bool

SparseBase.getfill — Function

getfill(A)

The value taken by all non-stored/implicit indices of A.

SparseBase.hasfixedsparsity — Method

hasfixedsparsity(::Type{A})::Bool

True if the sparsity pattern of the type A may be changed. A Diagonal type, for instance, may not have its sparsity pattern changed.

SparseBase.haszerobasedinternals — Method

haszerobasedinternals(::Type{A})::Bool

True if a type has zero based internal vectors.

This does not indicate that A is indexed in a zero-based manner, Instead, the internal representation is zero based. This is useful if C libraries implement much of the functionality of A, or A is frequently based to 0 based solvers.

SparseBase.isisovalued — Method

isisovalued(A)::Bool

True if a type contains a single value across all explicit indices.

SparseBase.isopaque — Method

isopaque(::Type{A})::Bool

True if internals may not be accessed directly. C owned types often set this to true.

SparseBase.issparse — Method

issparse(::T)::Bool
issparse(::DataType{T})::Bool

true if T is a sparse type. A sparse container is still considered sparse if all indices are explicitly stored.

SparseBase.mapstored! — Function

mapstored[!](f, [C], A)
mapstored(f, A ∪ B)
mapstored(f, (A ∩ B) ∪ C)

Map a function over the stored values of A(, B, C, ...).

Unions and intersections of A, B, C... operate elementwise over the unions and intersections of the patterns of those matrices.

TBD:

eadd in GraphBLAS doesn't work here, we need some notion of "passthrough".

missing doesn't work since missing + 2.0 = missing. GraphBLAS wants some ability to do NONSTORED + 2.0 = 2.0 for the eadd operation.

This could potentially lead to wrapper hell. I hope it doesn't, I think we would have

UnionPattern, IntersectionPattern (maybe ComplementPattern?). I don't think these wrappers are quite as bad as the ones in LinAlg, but I could be wrong.

Current implementations (SSGrB) only implement `mapstored!

SparseBase.nstored — Method

nstored(A)::Integer

Number of stored elements in A. In the dense case this is length(A)

SparseBase.setfill — Function

setfill!(A)::A
setfill(A)::B

Set the value taken by implicit indices of A to a new value. setfill produces a shallow copy of A with the new fill value.

SparseBase.storedindices — Function

storedindices(A)

An iterable over the stored indices of A. May be a direct view into internals, but is invalid to modify. May be a lazy iterator.

Returned indices should be a tuple (SoA) form.

SparseBase.storedvalues — Function

storedvalues(A)

An iterable over the stored values of A. May be a direct view into internals, but this is not a requirement, and so shouldn't be used to modify A.