SparseBase

Documentation for SparseBase.

SparseBase.CoordinateStore

SparseBase.UniformValue

SparseBase.convertinnerstore

SparseBase.deuniform!

SparseBase.filltype

SparseBase.findstored

SparseBase.getfill

SparseBase.hasfixedsparsity

SparseBase.iscoalesced

SparseBase.isopaque

SparseBase.isuniformvalued

SparseBase.mapstored!

SparseBase.npending

SparseBase.nstored

SparseBase.nzombies

SparseBase.setfill

SparseBase.storedindices

SparseBase.storedvalues

SparseBase.CoordinateStore — Type

CoordinateStore

SparseBase.UniformValue — Type

UniformValue{T}

Simple wrapper struct that denotes a uniform value for array input (all stored values are equal to a single value such as true).

SparseBase.convertinnerstore — Function

convertinnerstore(T::Type{<:AbstractSparseStore}, A)

Convert A into a similar format, with the internal sparse store converted to T.

SparseBase.deuniform! — Method

deuniform!(A)::A

If an array supports an iso representation convert it to the expected non-iso representation internally.

SparseBase.filltype — Method

filltype(A)

Type of implicit values of A. Most arrays either have no fill, or only support fill in the same domain as eltype(A).

SparseBase.findstored — Function

findstored(A)

An iterable over the stored indices and values of A. May be a direct view into internals, but this is not a requirement, and so shouldn't be used to modify A.

SparseBase.getfill — Function

getfill(A)

The value taken by all non-stored/implicit indices of A.

SparseBase.hasfixedsparsity — Method

hasfixedsparsity(::Type{A})::Bool

True if the sparsity pattern of the type A may be changed. A Diagonal type, for instance, may not have its sparsity pattern changed.

SparseBase.iscoalesced — Method

iscoalesced(::Type{A})::Bool = true

Whether each stored index is unique.

SparseBase.isopaque — Method

isopaque(::Type{A})::Bool

True if internals may not be accessed directly. C owned types set this to true.

SparseBase.isuniformvalued — Method

isuniformvalued(A)::Bool

True if a type contains a single value across all stored indices.

SparseBase.mapstored! — Function

mapstored[!](f, [E], [C], A)
mapstored(f, [E], A ∪ B)
mapstored(f, [E], (A ∩ B) ∪ C)

Map a function over the stored values of A(, B, C, ...).

Unions and intersections of A, B, C... operate elementwise over the unions and intersections of the patterns of those matrices.

Issues:

This could potentially lead to wrapper hell. I hope it doesn't, I think we would have

UnionPattern, IntersectionPattern (maybe ComplementPattern?). I don't think these wrappers are quite as bad as the ones in LinAlg, but I could be wrong.

Current implementations (SSGrB) only implement mapstored[!](f, [C], A), mapstored[!](f, [C], A ∪ B) and mapstored[!](f, [C], A ∩ B)

SparseBase.npending — Method

npending(A)::Integer

The number of pending insertions into A.

SparseBase.nstored — Method

nstored(A)::Integer

Number of stored elements in A. In the dense case this is length(A)

SparseBase.nzombies — Method

nzombies(A)::Integer

The number of pending deletions of values in A.

SparseBase.setfill — Function

setfill!(A, fill)::A
setfill(A, fill)::B

Set the value taken by implicit indices of A to a new value. setfill produces a shallow copy of A with the new fill value.

SparseBase.storedindices — Function

storedindices(A)

Iterables over the stored indices of A. May be a direct view into internals, but is invalid to modify. May be lazy iterables.

SparseBase.storedvalues — Function

storedvalues(A)

An iterable over the stored values of A. May be a direct view into internals, but this is not a requirement, and so shouldn't be used to modify A. Must be returned in the same order as storedindices.